+++ src/lib.rs
@@ -100,6 +100,7 @@ compile_error!("You can't enable `safe` and any `unsafe*` features at the same t
 pub mod code;
 pub mod data;
 pub mod error;
+pub mod gfx;
 pub mod mem;
 pub mod num;
 pub mod sys;
@@ -117,6 +118,7 @@ pub mod _all {
         code::all::*,
         data::all::*,
         error::all::*,
+        gfx::all::*,
         mem::all::*,
         num::all::*,
         sys::all::*,